+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Chg #161: Raise PHP version to 8.0 (@arogachev)
 - Enh #165: Improve perfomance (@arogachev)
 - Enh #134: Improve handling and control of `Assignment::$createdAt` (@arogachev)
+- Bug #178: Exclude parent role from `Manager::getAllChildRoles()` (@arogachev)
 
 ## 1.0.2 April 20, 2023